+++ projects/rocblas/library/src/hipblaslt_host.cpp
@@ -418,7 +418,8 @@ namespace
 template <typename Ti, typename To, typename Tc>
 rocblas_status runContractionProblemHipBlasLT(const RocblasContractionProblem<Ti, To, Tc>& prob,
                                               rocblas_gemm_algo                            algo,
-                                              int32_t solution_index)
+                                              int32_t  solution_index,
+                                              int32_t* selected_algo_index_out)
 {
     bool solution_query = algo == rocblas_gemm_algo_solution_index
                           && prob.flags & rocblas_gemm_flags_check_solution_index;
@@ -445,6 +446,9 @@ rocblas_status runContractionProblemHipBlasLT(const RocblasContractionProblem<Ti
         {
             return init;
         }
+
+        if(selected_algo_index_out)
+            *selected_algo_index_out = hipblaslt_ext::getIndexFromAlgo(heuristicResult.algo);
 
         auto gsu_malloc = prob.handle->gsu_malloc_by_size(workspace_size);
 
@@ -492,6 +496,9 @@ rocblas_status runContractionProblemHipBlasLT(const RocblasContractionProblem<Ti
         {
             return init;
         }
+
+        if(selected_algo_index_out)
+            *selected_algo_index_out = hipblaslt_ext::getIndexFromAlgo(heuristicResult.algo);
 
         auto gsu_malloc = prob.handle->gsu_malloc_by_size(workspace_size);
 
@@ -712,52 +719,64 @@ rocblas_status getAllSolutionsHipBlasLT(const RocblasContractionProblem<Ti, To, 
  ******************************************************************************/
 
 // Non-HPA/GEMM types
-template rocblas_status runContractionProblemHipBlasLT(
-    const RocblasContractionProblem<rocblas_half>&, rocblas_gemm_algo algo, int32_t solution_index);
+template rocblas_status
+    runContractionProblemHipBlasLT(const RocblasContractionProblem<rocblas_half>&,
+                                   rocblas_gemm_algo algo,
+                                   int32_t           solution_index,
+                                   int32_t*          selected_algo_index_out);
 
 template rocblas_status runContractionProblemHipBlasLT(const RocblasContractionProblem<float>&,
                                                        rocblas_gemm_algo algo,
-                                                       int32_t           solution_index);
+                                                       int32_t           solution_index,
+                                                       int32_t*          selected_algo_index_out);
 
 template rocblas_status runContractionProblemHipBlasLT(const RocblasContractionProblem<double>&,
                                                        rocblas_gemm_algo algo,
-                                                       int32_t           solution_index);
+                                                       int32_t           solution_index,
+                                                       int32_t*          selected_algo_index_out);
 
 template rocblas_status
     runContractionProblemHipBlasLT(const RocblasContractionProblem<rocblas_float_complex>&,
                                    rocblas_gemm_algo algo,
-                                   int32_t           solution_index);
+                                   int32_t           solution_index,
+                                   int32_t*          selected_algo_index_out);
 
 template rocblas_status
     runContractionProblemHipBlasLT(const RocblasContractionProblem<rocblas_double_complex>&,
                                    rocblas_gemm_algo algo,
-                                   int32_t           solution_index);
+                                   int32_t           solution_index,
+                                   int32_t*          selected_algo_index_out);
 
 // HPA types
 template rocblas_status runContractionProblemHipBlasLT(
     const RocblasContractionProblem<rocblas_half, rocblas_half, float>&,
     rocblas_gemm_algo algo,
-    int32_t           solution_index);
+    int32_t           solution_index,
+    int32_t*          selected_algo_index_out);
 
 template rocblas_status
     runContractionProblemHipBlasLT(const RocblasContractionProblem<rocblas_half, float, float>&,
                                    rocblas_gemm_algo algo,
-                                   int32_t           solution_index);
+                                   int32_t           solution_index,
+                                   int32_t*          selected_algo_index_out);
 
 template rocblas_status runContractionProblemHipBlasLT(
     const RocblasContractionProblem<rocblas_bfloat16, rocblas_bfloat16, float>&,
     rocblas_gemm_algo algo,
-    int32_t           solution_index);
+    int32_t           solution_index,
+    int32_t*          selected_algo_index_out);
 
 template rocblas_status
     runContractionProblemHipBlasLT(const RocblasContractionProblem<rocblas_bfloat16, float, float>&,
                                    rocblas_gemm_algo algo,
-                                   int32_t           solution_index);
+                                   int32_t           solution_index,
+                                   int32_t*          selected_algo_index_out);
 
 template rocblas_status
     runContractionProblemHipBlasLT(const RocblasContractionProblem<int8_t, int32_t, int32_t>&,
                                    rocblas_gemm_algo algo,
-                                   int32_t           solution_index);
+                                   int32_t           solution_index,
+                                   int32_t*          selected_algo_index_out);
 
 // ********** get all solutions explicits ********
 // Non-HPA/GEMM types